// youtube/rtk-query/lesson1/src/common/schemas/schemas.ts
import z from 'zod'
import { CurrentUserReaction } from '../enums'

export const tagSchema = z.object({
  id: z.string(),
  name: z.string(),
})

export const userSchema = z.object({
  id: z.string(),
  name: z.string(),
})

export const coverSchema = z.object({
  type: z.literal(['original', 'medium', 'thumbnail']),
  width: z.int().positive(),
  height: z.int().positive(),
  fileSize: z.int().positive(),
  url: z.url(),
})

export const imagesSchema = z.object({
  main: z.array(coverSchema),
})

export const currentUserReactionSchema = z.enum(CurrentUserReaction)